requirements.txt
streamlit
langchain
langchain-openai
openai
python-dotenv
pandas
sqlite-utils
tiktoken
jupyter
matplotlib